Address code review: add missing events to registry, fix minor issues (#233)
- Add 4 ComplexGateway domain events to EventTypeRegistryTests
- Document TryDeferredWorkflowCompletion as intentionally general-purpose
- Document ComplexGatewayJoinStateRemoved event usage pattern
- Fix ActivationCondition type: string.Empty → string? (consistent with ComplexGateway record)
- Move WaitForCondition helper to WorkflowTestBase (shared by Complex/Inclusive tests)
- Replace fragile Task.Delay with WaitForCondition polling in negative tests
