+++ src/Fleans/Fleans.Application.Tests/InclusiveGatewayTests.cs
@@ -264,28 +264,7 @@ public class InclusiveGatewayTests : WorkflowTestBase
 
     // --- Helper methods ---
 
-    /// <summary>
-    /// Polls the snapshot until the condition is met or times out.
-    /// Needed because condition evaluation happens asynchronously via Orleans streams.
-    /// </summary>
-    private async Task<InstanceStateSnapshot> WaitForCondition(
-        Guid instanceId, Func<InstanceStateSnapshot, bool> condition, int timeoutMs = 10000)
-    {
-        var deadline = DateTime.UtcNow.AddMilliseconds(timeoutMs);
-        while (DateTime.UtcNow < deadline)
-        {
-            var snapshot = await QueryService.GetStateSnapshot(instanceId);
-            if (snapshot != null && condition(snapshot))
-                return snapshot;
-            await Task.Delay(50);
-        }
 
-        var finalSnapshot = await QueryService.GetStateSnapshot(instanceId);
-        Assert.IsNotNull(finalSnapshot, "Snapshot was null after timeout");
-        Assert.IsTrue(condition(finalSnapshot),
-            $"Condition not met after {timeoutMs}ms. Active: [{string.Join(", ", finalSnapshot.ActiveActivities.Select(a => $"{a.ActivityId}({a.ActivityType})"))}], Completed: [{string.Join(", ", finalSnapshot.CompletedActivityIds)}], IsCompleted: {finalSnapshot.IsCompleted}");
-        return finalSnapshot;
-    }
 
     private static IWorkflowDefinition CreateInclusiveWorkflow(
         string task1Condition, string task2Condition, string task3Condition)
