+++ src/Fleans/Fleans.Persistence.Tests/EventTypeRegistryTests.cs
@@ -227,6 +227,10 @@ public class EventTypeRegistryTests
             (new GatewayForkCreated(id, null), nameof(GatewayForkCreated)),
             (new GatewayForkTokenAdded(id, id), nameof(GatewayForkTokenAdded)),
             (new GatewayForkRemoved(id), nameof(GatewayForkRemoved)),
+            (new ComplexGatewayJoinStateCreated("gw1", id, "_context._nroftoken >= 2", id), nameof(ComplexGatewayJoinStateCreated)),
+            (new ComplexGatewayJoinStateTokenIncremented("gw1"), nameof(ComplexGatewayJoinStateTokenIncremented)),
+            (new ComplexGatewayJoinStateFired("gw1"), nameof(ComplexGatewayJoinStateFired)),
+            (new ComplexGatewayJoinStateRemoved("gw1"), nameof(ComplexGatewayJoinStateRemoved)),
             (new ParentInfoSet(id, "parentAct"), nameof(ParentInfoSet)),
             (new ChildWorkflowLinked(id, id), nameof(ChildWorkflowLinked)),
             (new TimerCycleUpdated(id, "timer1", null), nameof(TimerCycleUpdated)),
